Okay custom emojis kinda work but need big time cleanup
They depend on the emoji picker but do not communicate that
the entire thing is super messy, named poorly and not extendable

The emoji picker should not provide any emojis.
There should be another emoji provider plugin that provides the discord
emojis. In the same format the url emoji provider provides them.